+++ EduTech/EduTech/src/main/java/com/EduTech/EduTech/repository/SoporteTecRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SoporteTecRepository extends JpaRepository<SoporteTec, Long> {
+public interface SoporteTecRepository extends JpaRepository<SoporteTec, Integer> {
 }
