+++ EduTech/EduTech/src/main/java/com/EduTech/EduTech/service/UsuarioService.java
@@ -19,7 +19,7 @@ public class UsuarioService {
         return usuarioRepository.findAll();
     }
 
-    public Usuario findById(long id) {
+    public Usuario findById(Integer id) {
         return usuarioRepository.findById(id).get();
     }
 
@@ -27,7 +27,7 @@ public class UsuarioService {
         return usuarioRepository.save(usuario);
     }
 
-    public void delete(Long id) {
+    public void delete(Integer id) {
         usuarioRepository.deleteById(id);
     }
 }
